lsu-pilot/pyproject.toml
[tool.poetry]
name = "lsu-pilot"
version = "0.1.0"
description = ""
authors = ["Donda <34602885+dondasasmita@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.6"
openai = "^1.52.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.8.0"
langchain = "^0.3.4"
pandas = "^2.2.3"
numpy = "^1.21"
scipy = "^1.14.1"
requests = "^2.32.3"
cairosvg = "^2.7.1"
celery = "^5.4.0"
redis = "^5.2.0"
flower = "^2.0.1"
pika = "^1.3.2"
debugpy = "^1.8.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
